+++ plugins/woocommerce-admin/packages/components/src/index.js
@@ -56,3 +56,4 @@ export { default as Tag } from './tag';
 export { default as TextControlWithAffixes } from './text-control-with-affixes';
 export { default as useFilters } from './higher-order/use-filters';
 export { default as ViewMoreList } from './view-more-list';
+export { default as WebPreview } from './web-preview';