+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add new component `<List />` for displaying interactive list items.
 - Fix z-index issue in `<Chart />` empty message.
 - Added a new `<SimpleSelectControl />` component.
+- Added a new `<WebPreview />` component.
 
 # 3.1.0
 - Added support for a countLabel prop on SearchListItem to allow custom counts.